Rename xsd.Schema to xsd.SchemaDocument and create one parent xsd.Schema
During parsing xsd imports are non transitive, however after the parsing
is done all global elements / types should be available even if it is
defined in a sub-sub schema. This is solved by making a clear
distinction between the main Schema interface and the various Schema
Documents its existing out of.