Fix issue #164 by using response namespace map
This code will use the xml namespace map provided in the SOAP response
from the server to fully expand the namespace prefixes used in the
xsi:type tags. This ensures that namespace prefixes used by the server
are expanded using the server generated xmlns tags rather than relying
on the request prefixes to remain the same.