Refactor the signature generation.
Never render nested global elements or types instead render
their prefixed name instead. This allows the removal of the depth
parameter. To get the proper prefixed name it is also now possible
to pass the schema to the signature() method so that the custom
prefix mapping is used.

The builtin xsd elements and types are now also handled as a
regular xsd:schema which allows some cleaning up