First attempt at adding support for elements which use a substitutionGroup.
This allows an any element to be substituted for any element which is within the first element's substitutionGroup.

This also overrides the keyname of the element in the kwargs struct, so that an substituted element may be referred to by its correct name.

This is tested and working on several of my trivial local tests. No real tests written yet.