[FIX] avoid error when the WSDL has a non binding port
This error can be replicated with this gist
https://gist.github.com/umiphos/543e92eef8c26aad44262bd55ba73baa
And this URL
https://e-factura.sunat.gob.pe/ol-ti-itcpfegem/billService?wsdl

You just make a new Client(URL) and you get the error from before.

Now, as you can see, that WSDL has a definition without bindings, that
makes an index error when zeep tries to get the services and bindings.
To fix that we apply this patch, in case of not bindings, we should just
warn about it and continue with the rest of the document.